+++ packages/askar/src/utils/askarStoreConfig.ts
@@ -48,8 +48,7 @@ export const uriFromStoreConfig = (
       database.credentials.password
     )}@${database.config.host}/${encodeURIComponent(storeConfig.id)}`
   } else {
-    // @ts-expect-error
-    throw new WalletError(`Storage type not supported: ${database.type}`)
+    throw new AskarError('Storage type not supported')
   }
 
   // Common config options