+++ packages/openid4vc/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.6.1
+
+### Patch Changes
+
+- 173ed6d: fix(openid4vc): add `grant_type_supported` to Credo hosted authorization server metadata
+- 8156e6e: fix: make `redirect_uri` required in the pushed authorization request endpoint (#2578)
+- 8156e6e: fix: also update authorization server configs on update of issuer record
+- a27de81: fix(openid4vc): loosen validation for multi-validation credential request. Some wallets send both `credential_configuration_id` and `format` which should not cause an error
+- Updated dependencies [9f60e1b]
+  - @credo-ts/core@0.6.1
+
 ## 0.6.0
 
 ### Minor Changes